made meal swipe an int

+++ api/models.py
@@ -14,7 +14,7 @@ class CustomUser(AbstractUser):
     The save() method is overridden to automatically update the user's meal swipe and flex dollar balances
     based on the associated MealPlanOption whenever the user is saved.
     """
-    meal_swipe_balance = models.DecimalField(max_digits=10, decimal_places=2, default=0.00)
+    meal_swipe_balance = models.IntegerField(default=0)
     flex_dollars = models.DecimalField(max_digits=10, decimal_places=2, default=0.00)
     meal_plan_option = models.ForeignKey('MealPlanOption', null=True, blank=True, on_delete=models.SET_NULL)
